data_collector/requirements.txt
requests
mysql-connector-python
yfinance
pandas